CyberOdyssey2024 Finals [Author]/WEB/WEBWEBWEB/backend/src/requirements.txt
asgiref==3.8.1
Django==4.2.17
Jinja2==3.1.4
django-pipeline==4.0.0
django-cors-headers==4.6.0
django-rest-framework==0.1.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
psycopg2-binary==2.9.9
PyJWT==2.10.1
rest-framework-simplejwt==0.0.2
sqlparse==0.5.2
typing-extensions==4.12.2
gunicorn==23.0.0
